@@ -4,19 +4,19 @@
   "description": "Chrome extension that enhances Z-Catering menu with food images",
   "type": "module",
   "scripts": {
-    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
-    "test:e2e": "node --experimental-vm-modules node_modules/jest/bin/jest.js -c jest-puppeteer.config.js tests/chrome.test.js",
-    "test:unit": "node --experimental-vm-modules node_modules/jest/bin/jest.js tests/dom-utils.test.js"
+    "test": "node --no-warnings --experimental-vm-modules node_modules/jest/bin/jest.js",
+    "test:e2e": "node --no-warnings --experimental-vm-modules node_modules/jest/bin/jest.js tests/chrome.test.js",
+    "test:unit": "node --no-warnings --experimental-vm-modules node_modules/jest/bin/jest.js tests/dom-utils.test.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.0",
-    "@babel/preset-env": "^7.24.0",
+    "@babel/core": "^7.27.1",
+    "@babel/preset-env": "^7.27.1",
     "jest": "^29.7.0",
-    "puppeteer": "^22.0.0",
-    "eslint": "^8.56.0",
-    "prettier": "^3.2.5",
+    "puppeteer": "^24.8.0",
+    "eslint": "^9.26.0",
+    "prettier": "^3.5.3",
     "cross-env": "^7.0.3",
-    "jest-puppeteer": "^9.0.2",
+    "jest-puppeteer": "^11.0.0",
     "jest-environment-jsdom": "^29.7.0"
   },
   "jest": {
