@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   launch: {
     headless: 'new',
     args: ['--no-sandbox', '--disable-setuid-sandbox']
